+++ src/telas/Home/componentes/Produtor.js
@@ -1,15 +1,29 @@
-import React, { useState } from "react";
+import React, { useReducer, useMemo } from "react";
 import { View, TouchableOpacity, Text, Image, StyleSheet } from 'react-native';
 
 import Estrelas from "../../../componentes/Estrelas";
 
 
+const distanciaEmMetros = (distancia) => {
+    console.log('distanciaEmMetros')
+    return `${distancia}m`;
+}
+
 export default function Produtor({nome, imagem, distancia, estrelas}) {
-    const [selecionado, setSelecionado ] = useState(false);
+    //const [selecionado, setSelecionado ] = useState(false);
+    const [selecionado, inverterSelecionado] = useReducer (
+        (selecionado) => !selecionado, 
+        false 
+    );
+
+    const distanciaTexto = useMemo(
+        () => distanciaEmMetros(distancia),
+        [distancia]
+        );
 
     return <TouchableOpacity 
             style={estilos.cartao}
-            onPress={() => setSelecionado(!selecionado)}
+            onPress={inverterSelecionado}
             >
         <Image style={estilos.imagem} source={imagem} accessibilityLabel={nome} />
         <View style={estilos.informacoes}>
@@ -21,7 +35,7 @@ export default function Produtor({nome, imagem, distancia, estrelas}) {
                 grande={selecionado}
                 />
             </View>
-            <Text style={estilos.distancia}>{distancia}</Text>
+            <Text style={estilos.distancia}>{distanciaTexto}</Text>
         </View>
     </TouchableOpacity>
 }
